Expanding readme.md
