+++ main.py
@@ -1,16 +1,36 @@
-# This is a sample Python script.
+import pandas as pd
 
-# Press Shift+F10 to execute it or replace it with your code.
-# Press Double Shift to search everywhere for classes, files, tool windows, actions, and settings.
+from ptm_hw1 import DateFinder
 
-
-def print_hi(name):
-    # Use a breakpoint in the code line below to debug your script.
-    print(f'Hi, {name}')  # Press Ctrl+F8 to toggle the breakpoint.
+"""
+This is the main function for the text mining project.  
 
 
-# Press the green button in the gutter to run the script.
-if __name__ == '__main__':
-    print_hi('PyCharm')
+"""
+if __name__ == "__main__":
 
-# See PyCharm help at https://www.jetbrains.com/help/pycharm/
+    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # read in the source file and convert to a series.
+    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    doc = []
+    with open(r'.\dates.txt') as file:
+        for line in file:
+            doc.append(line)
+
+    medical_dates_df = pd.Series(doc)
+
+    error_counter = 0
+
+    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Main Look runs the primary class on each text line, and in case of error, kicks out.
+    # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    for x in range(0, 500):
+        text_to_process = doc[x]
+        try:
+            found_date = DateFinder(raw_string=text_to_process)
+        except ValueError as uhoh:
+            print(f"count={x}\t{doc[x]}")
+            error_counter += 1
+
+    print(f"{error_counter=}")
+    print("fin")
