+++ text_mining_package/__init__.py
@@ -1,0 +1,1 @@
+from nic_date_date_class import NicDate
